+++ Properties/Settings.Designer.cs
@@ -37,18 +37,6 @@ namespace JeffRidder.NINA.Nightfront.Properties {
 
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("This is a notification")]
-        public string DefaultNotificationMessage {
-            get {
-                return ((string)(this["DefaultNotificationMessage"]));
-            }
-            set {
-                this["DefaultNotificationMessage"] = value;
-            }
-        }
-
-        [global::System.Configuration.UserScopedSettingAttribute()]
-        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
         [global::System.Configuration.DefaultSettingValueAttribute("")]
         public string NightFrontDataFolder {
             get {